First preliminary working version of USB drivers. Device Manager now accepts A, P and U input keys to print ATA, PCI and USB information respectively. Implemented RequestPages, IsPhysicalAddressMapped, IsVirtualAddressMapped, GetPhysicalAddress and GetVirtualAddress system calls. MemoryLayout can now handle merge/unmerge conflicts provided virtual addresses are mapped to the same physical addresses. USB drivers now work (aside from the Heap running out of memory). Dictionary now has additional Contains functions.

+++ Kernel/Libraries/Kernel.Hardware/BlockDevice.cs
@@ -86,7 +86,8 @@ namespace Kernel.Hardware
         /// <returns>The new byte array.</returns>
         public byte[] NewBlockArray(UInt32 aBlockCount)
         {
-            return new byte[aBlockCount * BlockSize];
+            //TODO: Support Conv_Ovf_I_Un IL op then remove the cast below
+            return new byte[aBlockCount * (uint)BlockSize];
         }
     }
 }
